Reset failed jobs instead of restarting them
How job arrays are handled now, it will be way easier to just reset
failed and cancelled jobs instead of actually restarting them. In the
end the effect will be more or less the same, and if we want to later,
we can add that this function also queues a new array.